Upgrade to BRCM SAI 10.0.0.3 EA
- Changes to upgrade to 10.0.0.3 EA Release and SDK 6.5.29
- Majority of the changes are to patches that hardcode the SDK version
  in the path.
- For the patches in CS00012296362.tar.gz (which are required to build
  kernel modules for 5.19), only 3 are required, the remaining patches
  have already been integrated into the SDK upstream.
- Most of the patch updates are exactly the same as the updates required
  for J3 bringup, please refer to
  https://docs.google.com/document/d/1r5LcP6SBUW0igLqo56XcwoIdimj_looeFLHON6uAymc
  for more details on the changes to these patches.

With the latest FBOSS OSS upstream pull, I could get some test cases in
HwVoqSwitchTest to only pass with the local build. I could not get the
test to pass with --known-good-hash.

I was able to load kernel 5.19 on QuartzDD manually and verify that the
kernel modules built for 5.19 with this changeset work.