Consume fboss-centos by its canonical name
fboss-centos repo has a canonical reponame of `code.arista.io/os/fboss-centos` but that when you consume here in arista-fboss it gets referenced to by its github name.
Ideally we would only use the canonical reponame of `code.arista.io/os/fboss-centos` to reference this thing.
This comes up in the context of BAR MUTs that try to roll new versions of fboss-centos into eos-trunk.  When we launch those MUTs, we get the repometadata for github.com/aristanetworks/fboss-centos which doesn't list any maintainers and so those BAR MUTs don't have any maintainers attached to them.