Fix bug in Qspf service, resulting in SEGFAULT when unsupported optics are plugged in.
When the plugged in optics are unsupported, qsfp service calculates the
start lane and hostlane count to both be 0, this results in code that
removes the port from transceiver programmed port collection.
The logic for removing the port from the collection was incorrect, we
were removing an item from a collection while iterating over it with a
range iterator.
We were seeing this failure on Vipers that had 100G LR4 optics in the
management port. FBOSS QSFP service only supports 100G CWDM4.