Make packetsnooper take buffer size to store more than one packet
Summary:
Currently, packet snopper has a single variable to store the last received packet.

The packet capture works as follows

* Create snooper -> send packet -> snooper wait

in the snooper,

* rx packet -> write to variable -> notify waiting thread

If more than one packet is received before the waiting thread processes the previous notify, its possible that we lose context on the earlier received packets.

In PTP tests, we  expect to receive 3 packets. However, the packets are received between the packet send and snooper wait and so only the last packet is received. We could solve this by waiting in a separate thread than the one that sends the packet but it will still not address a case where the waiting thread is busy with other processing while a notification comes in.

To properly solve this issue, created a buffer in packet snooper. The snooper will store all the packets received. The test case has to dequeue all the packets from the snooper before destroying it. If there are unhandled packets in the snooper when it is about to be destroyed, the snooper will fail the test. This will ensure that no spurious packets are received in the test

Reviewed By: maxwindiff

Differential Revision:
D61749023

Privacy Context Container: L1125642

fbshipit-source-id: 0ee813796afa46adaaeb726e741384b003001490