Meru800bia: Fix link test config
Remove most of the interfaces in the sw/interfaces section of the
Meru800bia link test config as these seem to be routed port configuration
and cause the link test to fail. Interface 11 is an inband recycle port
and needs to be retained for the test to pass.

Test Plan

Ran the `FbossOssLinkHwTest.py --sanity` test.