Method renamed
Summary:
As titled. This diff is a No-op.

But the following diffs in this stack will achieve the solution proposed in D74952681. That is, here is what we want to do to test NDP scale:
1. Fill ASIC with as close as possible to scale limit.
2. Then add more entries until we hit an error.
3. Confirm that number of entries successfully added is equal to ASIC claimed scale number.

X --> ASIC neigbor(NDP) scale number
B' --> Bulk entries added to SW switch state to program to ASIC.
B -->  kB' <= X,

Existing Logic:
```
1. Set B` = 8100 for TH4, 5100 for TH3, 512 for KODIAK3, 1000 for all other ASICs.
2. Add B' entries to switch state and apply the state to program the entries to ASIC
3. Now add more entries until we hit error as a result of INSUFFICIENT RESOURCE`
```

New Logic:
```
1. B' = MIN(1000, X)
2. loop = X/B';
    B = B' * loop;
    while(loop--):
        bulk add B' entries to SW switch state
        Apply switch state (program to ASIC)
3. Now we would have added B entries to ASIC, where B <= X
3. Now add more entries until we hit an error as a result of INSUFFICIENT RESOURCE
```

Benefits of new approach is, we don't need to define separate bulk programming number for every ASIC. And X%B' entries we add one by one. And need be we can adjust B' to 1024 as usually ASICs support 2^x scale.

Reviewed By: jasmeetbagga

Differential Revision: D75052739

fbshipit-source-id: 17bcebb83d7f952e18db23ca3f8bca673053eb93