TH3 NDP Scale Number
Summary:
Setting the right scale number for TH3.

BRCM claims this to be 8192, which indeed passed when we programmed 5100 entries in the initial state apply by setting 5100 as the return value from `bulkProgramCount()`, and then added the remaining entries 1 by 1.

But ideally we should be able to add all 8192 entries in the initial state apply and then add 1 more entry to hit the INSUFFICIENT error. Which is what we tries and it failed for TH3 and passed for TH4.

For TH3, the observation is, if we add greater than 5154 entries, test fails, if we set it to less than or equal to 5154 it passes.

**Theoretical Analysis:**

I think it could be possible that, SDK doesn't get chance to re-hash the neighbor entries and de-fragment the hash table in ASIC if we add more than 5154 entries. prasoon will create a CSP and work with BRCM. Once we receive the SDK fix, we will update the number as per BRCM recommendation.

Reviewed By: zechengh09

Differential Revision:
D75211129

Privacy Context Container: L1125642

fbshipit-source-id: cfebe270438923fe65adcbdf13f6be2f506df790