bfg: Update ccache URL to include password
We cannot allow un-authenticated access to the redis instance
backing fboss's ccache. In this change
https://github.com/aristanetworks/barney-ops/pull/1211 I created
a password and configured bsy to inject it into build environments
as a secret. This should allow the bsy builds to authenticate to
the redis instance once it gets restarted to use a password.

This means that older build environments which try to access redis
without a password will not be able to access redis. This does not
cause the build to fail, it just behaves as a cache miss.

As of right now, we don't have a good way of providing BUS users
the ability to read from the cache, but I'm hoping to restore that
functionality in the near future by providing a set of readonly
credentials that can be distributed to busses.