barneyfile/bus: Bake CCACHE_SECONDARY_STORAGE into bus environment
Host-settings are not available on BBM so instead, for now, let's
just bake the ccache remote storage url into the bus image. This
host-setting is no longer a secret because we now get it from WSS.

By baking it into the image itself (rather than the entry settings),
we can make it so that users have CCACHE_SECONDARY_STORAGE properly
set up when they enter their bus images.